+++ CHANGELOG.md
@@ -6,9 +6,10 @@ The format is roughly based on [Keep a Changelog](https://keepachangelog.com/en/
 
 <!-- start-of-changelog -->
 
-## Work in progress: 0.50.0
+## Work in progress: 0.50
 
 - Harden cross-backend splitting for cases without `load_collection` nodes [#202](https://github.com/Open-EO/openeo-aggregator/issues/202)
+- Partitioned job tracking: avoid getting stuck in "running" state when upstream job fails to start ([#203](https://github.com/Open-EO/openeo-aggregator/issues/203))
 
 
 ## 0.49.0